Finish the upper section of the page (probably)
Added the background image to the upper part of the page, will possibly
need to make it scroll along with the page. Added stylings to header,
text and image. Header is still bad but there is nothing I can do at
this point. Borders seem fine, need to work a little on the button.